internal/sqlsmith: pre-split and scatter ranges
This commit modifies `seed` and `seed-multi-region` sqlsmith setups to
manually split the seed table and then scatter the ranges. This should
make the setups more interesting.

This required allowing secondary tenants to run these commands in some
tests that use these setups.

Release note: None
